requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
dash==2.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.2
Pillow==9.0.1
plotly==5.5.0
requests==2.27.1
six==1.16.0
tenacity==8.0.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu 
torchvision==0.10.0+cpu
